pyproject.toml
[project]
name = "labelable"
version = "0.1.1.dev14"
description = "A general purpose label printing API and basic UI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "fastui>=0.7.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "pyserial>=3.5",
    "pillow>=11.0.0",
    "python-multipart>=0.0.9",
    "aiohttp>=3.9.0",
    "qrcode[pil]>=7.4.0",
    "pylibdmtx>=0.1.10",
    "setuptools>=80.10.2",
    "fonttools>=4.61.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0,<=0.23.7",  # Pinned for HA wheels compatibility
    "pytest-cov>=4.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[project.scripts]
labelable = "labelable.__main__:main"
labelable-render = "labelable.cli.render:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/labelable"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[dependency-groups]
dev = [
    "basedpyright>=1.37.4",
]

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
include = ["src"]
exclude = ["**/__pycache__", "**/node_modules"]
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
